Launch pajam shiny app

Launch pajam shiny app

launch_pajam(
  ...,
  width = 2200,
  port = 8888,
  host = "0.0.0.0",
  options = list(width = width, host = host, port = port)
)

Arguments

...
additional arguments are ignored.

width
numeric number of pixels to define the default page size. This value is useful in allowing sufficient width to the main heatmap, even though the heatmap can be resized by dragging the small triangle on the bottom-right corner of the heatmap.

port
numeric value indicating the server port used for the R-shiny server. The default 8888 allows R-shiny to run under linux-alike without requiring root privilege.

host
character indicating the host IP address, or host server name. This value also restricts all incoming requests to match this value, except when host="0.0.0.0" in which case it will accept all incoming requests.

options
list of additional R-shiny server options.

See also